move admin lecture transcription to core module
