+++ src/main/webapp/app/core/admin/admin.routes.ts
@@ -163,7 +163,8 @@ const routes: Routes = [
     },
     {
         path: 'lecture-transcription-ingestion',
-        loadComponent: () => import('app/admin/lecture-transcription-ingestion/lecture-transcription-ingestion.component').then((m) => m.LectureTranscriptionIngestionComponent),
+        loadComponent: () =>
+            import('app/core/admin/lecture-transcription-ingestion/lecture-transcription-ingestion.component').then((m) => m.LectureTranscriptionIngestionComponent),
         data: {
             authorities: [Authority.ADMIN],
             pageTitle: 'global.menu.admin.lectureTranscription',
