+++ src/main/webapp/app/core/admin/lecture-transcription-ingestion/lecture-transcription-ingestion.component.scss
