+++ src/main/webapp/app/core/admin/lecture-transcription-ingestion/lecture-transcription-ingestion.component.ts
@@ -1,7 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { faCheck } from '@fortawesome/free-solid-svg-icons';
 import { FormsModule } from '@angular/forms';
-import { LectureTranscriptionService } from 'app/admin/lecture-transcription-ingestion/service/lecture-transcription.service';
+import { LectureTranscriptionService } from 'app/core/admin/lecture-transcription-ingestion/service/lecture-transcription.service';
 import { AlertService } from 'app/shared/service/alert.service';
 import { ButtonComponent } from 'app/shared/components/buttons/button/button.component';
