+++ src/main/webapp/app/core/admin/lecture-transcription-ingestion/service/lecture-transcription.service.ts
